Add files via upload
Diabetes is a common disease that is incurable and fatal. Millions of people worldwide have diabetes and it directly affects people’s
lives. Early diagnosis helps reduce the effects of diabetes and improve the life quality of patients, but in common case people live with
diabetes for years before getting diagnosed. Early diagnosis can be done by applying machine learning methods on existing data of
patients.